+++ server/src/main/java/io/anemos/metastore/RegistryService.java
@@ -108,7 +108,7 @@ public class RegistryService extends RegistryGrpc.RegistryImplBase {
                   .asRuntimeException());
           return;
         }
-        registry.update(registry.ref(), in, report);
+        registry.update(registry.ref(), in, report, request.getComment());
       }
 
       responseObserver.onNext(